refactor(web): move attachment/DOCX adapters out of use-ws-runtime
use-ws-runtime.ts held attachment/DOCX conversion adapters
(CodeTextAttachmentAdapter, OfficeDocumentAttachmentAdapter,
normalizeDocxTableHtml, escapeXmlAttribute, attachmentAdapter) that have
nothing to do with the WS runtime's socket state machine — quick-win
extraction called out in #1074, ahead of the larger ChatSocket reducer
extraction which stays out of scope for this change.

Behaviour-preserving move: src/lib/attachment-adapters.ts now owns the
adapters, and use-ws-runtime.ts imports attachmentAdapter from it — and
loses the CLIENT_MAX_ATTACHMENT_SIZE_BYTES import that had been unused
since #1087.

Both rebase conflicts landed on the same decision, and it is worth
recording which way they were resolved. This branch forked before #1087
collapsed `OfficeDocumentAttachmentAdapter.add()` onto the shared
`oversizeAttachmentError` helper, so the moved code keeps MAIN's version,
not this branch's hand-rolled size check. That is not a formality: the
two long-standing size tests match `/too large/i` and `/huge\.docx/`, so
BOTH wordings satisfy them, and resolving the other way would have
reinstated a second sentence for one limit with a green suite. Main's own
guard against exactly that (75896d078) is carried through the rename
rather than duplicated with this branch's equivalent.

Tests moved, none deleted: attachment-routing.test.ts and
office-document-adapter.test.ts merge into
src/__tests__/lib/attachment-adapters.test.ts, and the
normalize-docx-table-html-drift.test.ts guard follows the code it guards
into __tests__/lib/ with its path updated.

Docs-not-needed: internal refactor, code moved between modules only
